update graph implementation of controllers
